+++ stroom-explorer-server/src/main/java/stroom/explorer/server/ExplorerNodeServiceImpl.java
@@ -1,5 +1,7 @@
 package stroom.explorer.server;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 import stroom.entity.shared.PermissionInheritance;
 import stroom.explorer.shared.ExplorerConstants;
@@ -19,6 +21,8 @@ import java.util.stream.Collectors;
 
 @Component
 class ExplorerNodeServiceImpl implements ExplorerNodeService {
+    private static final Logger LOGGER = LoggerFactory.getLogger(ExplorerNodeServiceImpl.class);
+
     // TODO : This is a temporary means to set tags on nodes for the purpose of finding data source nodes.
     // TODO : The explorer will eventually allow a user to set custom tags and to find nodes searching by tag.
     private static Map<String, String> DEFAULT_TAG_MAP = new HashMap<>();
@@ -43,13 +47,30 @@ class ExplorerNodeServiceImpl implements ExplorerNodeService {
 
     @Override
     public void createNode(final DocRef docRef, final DocRef destinationFolderRef, final PermissionInheritance permissionInheritance) {
+        // Ensure permission inheritance is set to something.
+        PermissionInheritance perms = permissionInheritance;
+        if (perms == null) {
+            LOGGER.warn("Null permission inheritance supplied for create operation");
+            perms = PermissionInheritance.DESTINATION;
+        }
+
         // Set the permissions on the new item.
-        if (permissionInheritance == null || PermissionInheritance.NONE.equals(permissionInheritance)) {
-            // Make the new item owned by the current user.
-            addDocumentPermissions(null, docRef, true);
-        } else if (PermissionInheritance.COMBINED.equals(permissionInheritance) || PermissionInheritance.INHERIT.equals(permissionInheritance)) {
-            // Copy permissions from the containing folder and make the new item owned by the current user.
-            addDocumentPermissions(destinationFolderRef, docRef, true);
+        try {
+            switch (perms) {
+                case NONE:
+                    // Make the new item owned by the current user.
+                    addDocumentPermissions(null, docRef, true);
+                    break;
+                case DESTINATION:
+                    // Copy permissions from the containing folder and make the new item owned by the current user.
+                    addDocumentPermissions(destinationFolderRef, docRef, true);
+                    break;
+                default:
+                    LOGGER.error("Unexpected permission inheritance '" + perms + "' supplied for create operation");
+                    break;
+            }
+        } catch (final Exception e) {
+            LOGGER.error(e.getMessage(), e);
         }
 
         addNode(destinationFolderRef, docRef);
@@ -57,18 +78,39 @@ class ExplorerNodeServiceImpl implements ExplorerNodeService {
 
     @Override
     public void copyNode(final DocRef sourceDocRef, final DocRef destDocRef, final DocRef destinationFolderRef, final PermissionInheritance permissionInheritance) {
+        // Ensure permission inheritance is set to something.
+        PermissionInheritance perms = permissionInheritance;
+        if (perms == null) {
+            LOGGER.warn("Null permission inheritance supplied for copy operation");
+            perms = PermissionInheritance.DESTINATION;
+        }
+
         // Set the permissions on the copied item.
-        if (permissionInheritance == null || PermissionInheritance.NONE.equals(permissionInheritance)) {
-            // Copy permissions from the original and make the new item owned by the current user.
-            addDocumentPermissions(sourceDocRef, destDocRef, true);
-        } else if (PermissionInheritance.COMBINED.equals(permissionInheritance)) {
-            // Copy permissions from the original and make the new item owned by the current user.
-            addDocumentPermissions(sourceDocRef, destDocRef, true);
-            // Add additional permissions from the folder of the new copy.
-            addDocumentPermissions(destinationFolderRef, destDocRef, true);
-        } else if (PermissionInheritance.INHERIT.equals(permissionInheritance)) {
-            // Just add permissions from the folder of the new copy and make the new item owned by the current user.
-            addDocumentPermissions(destinationFolderRef, destDocRef, true);
+        try {
+            switch (perms) {
+                case NONE:
+                    // Ignore original permissions, ignore permissions of the destination folder, just make the new item owned by the current user.
+                    addDocumentPermissions(null, destDocRef, true);
+                    break;
+                case SOURCE:
+                    // Copy permissions from the original, ignore permissions of the destination folder, and make the new item owned by the current user.
+                    addDocumentPermissions(sourceDocRef, destDocRef, true);
+                    break;
+                case DESTINATION:
+                    // Ignore permissions of the original, add permissions of the destination folder, and make the new item owned by the current user.
+                    addDocumentPermissions(destinationFolderRef, destDocRef, true);
+                    break;
+                case COMBINED:
+                    // Copy permissions from the original, add permissions of the destination folder, and make the new item owned by the current user.
+                    addDocumentPermissions(sourceDocRef, destDocRef, true);
+                    addDocumentPermissions(destinationFolderRef, destDocRef, true);
+                    break;
+                default:
+                    LOGGER.error("Unexpected permission inheritance '" + perms + "' supplied for copy operation");
+                    break;
+            }
+        } catch (final Exception e) {
+            LOGGER.error(e.getMessage(), e);
         }
 
         addNode(destinationFolderRef, destDocRef);
@@ -76,17 +118,39 @@ class ExplorerNodeServiceImpl implements ExplorerNodeService {
 
     @Override
     public void moveNode(final DocRef docRef, final DocRef destinationFolderRef, final PermissionInheritance permissionInheritance) {
+        // Ensure permission inheritance is set to something.
+        PermissionInheritance perms = permissionInheritance;
+        if (perms == null) {
+            LOGGER.warn("Null permission inheritance supplied for copy operation");
+            perms = PermissionInheritance.DESTINATION;
+        }
+
         // Set the permissions on the moved item.
-        if (permissionInheritance != null) {
-            if (PermissionInheritance.COMBINED.equals(permissionInheritance)) {
-                // Add permissions from the new folder.
-                addDocumentPermissions(destinationFolderRef, docRef, false);
-            } else if (PermissionInheritance.INHERIT.equals(permissionInheritance)) {
-                // Clear existing permissions from this item.
-                securityContext.clearDocumentPermissions(docRef.getType(), docRef.getUuid());
-                // Add permissions from the new folder and make the current user the owner.
-                addDocumentPermissions(destinationFolderRef, docRef, true);
+        try {
+            switch (perms) {
+                case NONE:
+                    // Remove all current permissions, ignore permissions of the destination folder, just make the new item owned by the current user.
+                    clearDocumentPermissions(docRef);
+                    addDocumentPermissions(null, docRef, true);
+                    break;
+                case SOURCE:
+                    // We are keeping the permissions that we already have so do nothing.
+                    break;
+                case DESTINATION:
+                    // Remove all current permissions, add permissions of the destination folder, and make the new item owned by the current user.
+                    clearDocumentPermissions(docRef);
+                    addDocumentPermissions(destinationFolderRef, docRef, true);
+                    break;
+                case COMBINED:
+                    // Keep all current permissions, add permissions of the destination folder, and make the new item owned by the current user.
+                    addDocumentPermissions(destinationFolderRef, docRef, true);
+                    break;
+                default:
+                    LOGGER.error("Unexpected permission inheritance '" + perms + "' supplied for move operation");
+                    break;
             }
+        } catch (final Exception e) {
+            LOGGER.error(e.getMessage(), e);
         }
 
         moveNode(destinationFolderRef, docRef);
@@ -276,6 +340,10 @@ class ExplorerNodeServiceImpl implements ExplorerNodeService {
         }
 
         securityContext.addDocumentPermissions(sourceType, sourceUuid, destType, destUuid, owner);
+    }
+
+    private void clearDocumentPermissions(final DocRef docRef) {
+        securityContext.clearDocumentPermissions(docRef.getType(), docRef.getUuid());
     }
 
     private ExplorerNode createExplorerNode(final ExplorerTreeNode explorerTreeNode) {